Add files via upload
in file part1, We deal with audio signal processing. We draw its spectrogram, Using Fourier series and high-pass and low-pass filters, we eliminate noise.
in file part2, We apply two-dimensional convolution on the image and apply some filters on the image.
in all parts, we draw the pictures and charts.